requirements.txt
streamlit
pandas
openai>=1.0.0
python-dotenv